// src/types/EmployeeType.ts
export type EmployeeType = {
  id: string;
  employeeCode: string;
  name: string;
  fatherName: string;
  cnic: string;
  dob: string;
  email: string;
  profile: string;
  designation: string;
  mobile: string;
  emergencyContact: string;
  dateOfConfirmation: string;
  dateOfJoining: string;
  currentAddress: string;
  permanentAddress: string;
  niNumber: string;
  designationId: string;
};
export type NewEmployeeType = {
  FullName: string;
  FatherName: string;
  CNIC: string;
  DOB: Date;
  ProfilePictureFile: string;
  PersonalEmail: string;
  OfficialEmail: string;
  NiNumber: string;
  Email: string;
  DesignationId: string;
  Mobile: string;
  EmergencyContact: string;
  DateOfConfirmation: Date;
  DateOfJoining: Date;
  CurrentAddress: string;
  PermanentAddress: string;
  Username: string;
  Password: string;
  Role: string;
};